Remove legacy logic and refactor for new
