+++ h-util-ui/Electron/operations/handler.helpers.ts
@@ -1,7 +1,6 @@
 import path from 'path';
 import fs from 'fs/promises';
 
-import { withTimer, promises } from '@common/common';
 import {
     splitFileNameFromPath,
     checkSupportedExt,
@@ -12,108 +11,8 @@ import {
 } from '@common/fileops';
 import { slugify } from '@shared/common.utils';
 
-import { CommonContext, FileOptions, FileWithMeta, ModuleHandler, ModuleOptions } from '@util/types';
-import { ActionModule, ProcessingModule, ProcessingModuleType } from '@shared/common.types';
-import output from '@util/output';
+import { FileOptions, FileWithMeta, ModuleOptions } from '@util/types';
 import { ExtraData, RenameTemplates } from '@shared/common.constants';
-import { MODULE_MAP } from './modules/moduleMap';
-import { ProcessingError } from '@util/errors';
-
-type WithFileListHandlingArgs<T extends object> = {
-    fileOptions: FileOptions;
-    moduleHandler: ModuleHandler<T>;
-    /** Information shared across modules if necessary */
-    context?: T;
-    onProgress?: (label: string, progress: number) => void;
-    /** Configurations passed from the client: module configs */
-    clientOptions?: ActionModule['options'];
-    onEvent?: (message: string) => void;
-};
-
-/**
- * Take a module and iterate through with its handler over files
- */
-export const withFileListHandling = async <T extends object = {}>({
-    fileOptions,
-    moduleHandler,
-    onProgress,
-    context,
-    clientOptions,
-}: WithFileListHandlingArgs<T>) => {
-    const { filter, handler } = moduleHandler;
-
-    let processed = 0;
-    let filtered = 0;
-    let errored = 0;
-    let timeTaken = 0;
-
-    /** For handlers to persist any temporary data needed */
-    const dataStore: Record<string, any> = {};
-
-    if (handler)
-        await withTimer(
-            async () => {
-                await promises.map(fileOptions.filesWithMeta, async (fileWithMeta, i) => {
-                    let topFileName = 'Unknown file';
-                    try {
-                        const { filePath } = fileWithMeta;
-
-                        const { fileName } = splitFileNameFromPath(filePath);
-                        topFileName = fileName;
-
-                        const shouldHandle = filter ? await filter(filePath) : true;
-
-                        onProgress?.(fileName, Math.ceil((i / fileOptions.filesWithMeta.length) * 100));
-
-                        /**
-                         * If this file has been previously marked to skip, or we determine that now
-                         */
-                        if (!shouldHandle || (clientOptions?.skipPreviouslyFailed && fileWithMeta.previouslySkipped)) {
-                            filtered++;
-                            if (clientOptions?.skipPreviouslyFailed) fileWithMeta.previouslySkipped = true;
-                            return;
-                        }
-
-                        await handler(
-                            fileWithMeta,
-                            {
-                                onSuccess: () => {},
-                                context,
-                                clientOptions,
-                            },
-                            dataStore,
-                        );
-
-                        processed++;
-                    } catch (e: any) {
-                        console.log('Error with handler:', e);
-                        addEventLogForReport({ context }, topFileName, 'errored', e.message);
-
-                        errored++;
-
-                        if (!clientOptions?.ignoreErrors) throw e;
-                    }
-                });
-
-                await moduleHandler.onDone?.({ clientOptions, context }, dataStore, fileOptions);
-            },
-            (time) => {
-                timeTaken = time;
-            },
-        );
-
-    if (!handler && moduleHandler.onDone)
-        await moduleHandler.onDone({ clientOptions, context }, dataStore, fileOptions);
-
-    output.log(`Module ran in ${timeTaken}ms, ${fileOptions.filesWithMeta.length} files`);
-
-    return {
-        timeTaken,
-        errored,
-        filtered,
-        processed,
-    };
-};
 
 export const addEventLogForReport = (
     opts: Partial<ModuleOptions<{}>>,
@@ -219,50 +118,5 @@ export const populateDataDict = async ({
             return;
         default:
             return;
-    }
-};
-
-type RunProcessingModuleOpts = {
-    onBeforeRun?: () => void;
-    onProgress?: (label: string, progress: number) => void;
-    commonContext?: CommonContext;
-};
-
-export const runProcessingModule = async (
-    processingModule: ProcessingModule,
-    fileOptions: FileOptions,
-    opts: RunProcessingModuleOpts,
-) => {
-    const { onBeforeRun, onProgress, commonContext } = opts;
-
-    if (processingModule.type === ProcessingModuleType.branch) throw new Error('Branch not supported yet');
-
-    const moduleHandler = MODULE_MAP[processingModule.type];
-
-    onBeforeRun?.();
-
-    if (!moduleHandler) {
-        output.log(`Module ${processingModule.type} not yet supported`);
-        return;
-    }
-
-    output.log(`Processing module ${processingModule.type}`);
-
-    try {
-        /** One module's processing */
-        await withFileListHandling({
-            fileOptions,
-            clientOptions: processingModule.options,
-            moduleHandler,
-            onProgress,
-            context: commonContext,
-        });
-    } catch (e) {
-        if (e instanceof ProcessingError) {
-            // in future we may ignore
-        } else {
-            console.error('[runPipeline]', e);
-            throw e;
-        }
     }
 };
